pyproject.toml
[project]
name = "synapse-tool"
version = "0.1.0"
description = "AI Context Augmentation Tool for Large Scale Codebases"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.7.0",
    "tree-sitter>=0.24.0",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.3.1",
    "networkx>=3.2.1",
    "pydantic>=2.6.0",
    "watchdog>=4.0.0",
    "tree-sitter-python>=0.24.0",
    "tree-sitter-javascript>=0.24.0",
    "tree-sitter-typescript>=0.23.0",
    "tree-sitter-go>=0.24.0",
    "tree-sitter-java>=0.23.0",
    "tree-sitter-cpp>=0.23.0",
    "tree-sitter-c>=0.24.0",
    "tree-sitter-rust>=0.24.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/synapse", "src/synapse/templates"]

[project.scripts]
synapse = "synapse.cli:app"
